fix: potential race condition when releasing environment (#37)
Currently, when an environment is acquired, the allocation it was
assigned to is not stored in the `environments` table. Therefore,
operating on an environment does not require the allocation id. Since
only a single allocation can acquire an environment, this is ok **as
long as** we can guarantee that there are no concurrent operations on
the same environment.

However, this is not so easy to guarantee, and also goes against the
design, which favors idempotancy over coordination.

To give a concrete example of the problem, imagine the following
timeline:

```console
t=0: environment A is acquired by test1.
t=1: environment A is released and the cleanup timeout event is scheduled for t + 3.
t=2: environment A is successfully cleaned and released.
t=3: environment A is acquired by test2.
t=4: cleanup timeout event is triggered and marks the environment A as `dirty`  <-- Whoops. No good.
```

To avoid this, the allocation id is now stored in the `environments`
table as well. This allows operations to detect when they are operating
on a reallocated environment, and act accordingly.

Allocation ID generation moved to the API layer so that it can be shared
across multiple tables (and avoid `INSERT->INSERT->UPDATE`)