chore: stack outputs to auto discover service attributes in integration tests (#40)
Currently, the integration test session discovers service attributes
(e.g the `environments` table name) using a lookup based on its expected
name pattern. This approach is brittle because it relies on specific
strings to exist in the name, even though the name is auto generated by
CFN. It also means a special lookup function is required every time we
add an attribute the tests might need knowledge of.

Instead, the test now creates stack outputs for each of the service
attributes required by tests. The session will then describe the stack
and fetch its outputs to determine the value of an attribute.

Also this PR formalizes service environment variables as a concrete
type, so that reading values is done in a type safe way.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>