fix: environments can be released even though they shouldn't be (#50)
Currently, there is no restriction on the status an environment should
be in before releasing it. This isn't great because our service has a
clear state transition expectation: `in-use -> cleaning -> released`.
Any other transition should be denied by the DB and explicitly handled
by the caller, based on the thrown exception.

For example, it might be that the cleanup task finishes its execution
AFTER the timeout has already expired.

> It really shouldn't because the timeout event is scheduled to trigger
only after the cleanup has finished. But its not a strict guarantee,
various delays can happen and its better that our functions, if they
can, don't rely solely on sequencing.

In this case, the timeout event marks the environment as `dirty` as
expected, but than the cleanup task ends up releasing the environment.
This being a problem is debatable, because presumably the environment is
indeed clean and can be restored back to the service. However, a dirty
environment triggers an alarm, so it would be strange for an operator to
get an alarm about a `dirty` environment, and see that it was actually
already released.

This scenario should be pretty rare, I don't expect many such "false"
alarms. We can always schedule the cleanup timeout even further from the
cleanup duration if we see they are too close.

### Also in this PR

- Formalize environment status with an enum.