chore: log allocation id everywhere (#59)
All service operations are executed for a specific allocation, lets make
sure we log that allocation in every log line. This will allow us to
formulate a LogInsights query that retrieves allocation specific logs
for all operations.

In the future we might want to do something with Xray or structured
logging, but for now this is easy enough and gets us pretty far.