fix: cannot retry cleanup task on dirty environment (#78)
Currently, an environment can only be released if its status is
`cleaning`, because this is the status the cleanup task should expect
when it first runs on an environment.

However, this prevents us from retrying the task on human intervention
of dirty environments. Instead, lets just assert that the status is not
`in-use`.