feat: custom domain (#84)
Optionally create a custom domain for the API endpoint if a hosted zone
is provided.

- I'm still trying to figure out how to delegate subdomains, so for now,
the name of the custom domain will be the root domain of the hosted
zone.
- Integration tests don't use a custom domain for now because it would
require a pre-configured hosted zone to exist in each of our accounts.

> I think that eventually we would run the integration tests of this
service, using an already deployed version of this exact service. This
will allow us to ask the service@v1 for an environment to deploy and
test service@v2, and that environment can contain a hosted zone :)

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>