feat: respond with role arn instead of credentials (#97)
AppSec didn't like the fact we are directly returning credentials in
response to an allocation request. I see their point.
So instead, we will only return a role ARN that the caller would have to
assume in order to operate on the allocated environment.

This makes the environment registration process somewhat more involved
because the role would need to allow the caller to assume it, but it
does provide better security and protection against account takeover.