fix: default endpoint is enabled by default (#101)
Currently, the default `execute-api` endpoint is enabled by default, and
is only disabled in case a hosted zone is provided (in which case the
service should be accessed via the custom domain).

This however makes it a little too easy to make a mistake during
production deployment by not remembering to pass a `hostedZone`, and
thus enabling the default endpoint.

We actually only need the default endpoint in integration tests, so lets
always disable the default endpoint, and escape hatch our way around it
in integration tests alone. That way, a standard instantiation of the
service will result in a disabled endpoint, which is what we want.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>